requirements.txt
# Core
Django==6.0.2
psycopg[binary]==3.2.6
daphne==4.1.2

# Auth
django-allauth[socialaccount]==65.4.1

# AI / LLM
openai==1.68.2

# Document processing
PyMuPDF==1.25.3
python-docx==1.1.2

# Async tasks
celery==5.4.0
redis==5.2.1

# Google Drive
google-api-python-client==2.159.0
google-auth==2.37.0

# Utilities
python-dotenv==1.0.1
